requirements.txt
setuptools~=57.4.0
numpy~=1.19.5
twython~=3.9.1
requests~=2.26.0
Pillow~=8.3.1
scikit-image~=0.18.1
pandas~=1.3.0
imutils~=0.5.3
gpiozero~=1.6.2
picamera
matplotlib
scipy~=1.7.0